Handling request for different namespace
