+++ src/data/methods/EIP155Data.methods.ts
@@ -1,0 +1,18 @@
+/**
+ * @desc Refference list of eip155 chains
+ * @url https://chainlist.org
+ */
+
+/**
+ * Methods
+ */
+export const EIP155_SIGNING_METHODS = {
+	PERSONAL_SIGN: 'personal_sign',
+	ETH_SIGN: 'eth_sign',
+	ETH_SIGN_TRANSACTION: 'eth_signTransaction',
+	ETH_SIGN_TYPED_DATA: 'eth_signTypedData',
+	ETH_SIGN_TYPED_DATA_V3: 'eth_signTypedData_v3',
+	ETH_SIGN_TYPED_DATA_V4: 'eth_signTypedData_v4',
+	ETH_SEND_TRANSACTION: 'eth_sendTransaction',
+	SWITCH_CHAIN: 'wallet_switchEthereumChain',
+}
