+++ src/data/methods/SolanaData.methods.ts
